refactoring: make type conversions more explicit in JSON parsing
It is now more obvious when precision between a value from the
JSON is different from the stored value.

+++ src/json/NLohmannJsonOwm.cpp
@@ -46,14 +46,14 @@ bool NLohmannJsonOwm::parseSingleWeatherItem(const value_type& value, Weather& w
       {
         weather.setTemperatureKelvin(kelvinRounded);
       }
-      weather.setTemperatureCelsius(kelvinRaw - 273.15);
+      weather.setTemperatureCelsius(kelvinRaw - 273.15f);
       // Avoid values like 6.9999... ° C by rounding, if appropriate.
       const float celsiusRounded = std::round(weather.temperatureCelsius());
       if (std::fabs(celsiusRounded - weather.temperatureCelsius()) < 0.005)
       {
         weather.setTemperatureCelsius(celsiusRounded);
       }
-      weather.setTemperatureFahrenheit((kelvinRaw - 273.15) * 1.8 + 32.0f);
+      weather.setTemperatureFahrenheit(static_cast<float>((kelvinRaw - 273.15f) * 1.8 + 32.0));
       // Avoid values like 6.9999... ° F by rounding, if appropriate.
       const float fahrenheitRounded = std::round(weather.temperatureFahrenheit());
       if (std::fabs(fahrenheitRounded - weather.temperatureFahrenheit()) < 0.005)
@@ -66,7 +66,7 @@ bool NLohmannJsonOwm::parseSingleWeatherItem(const value_type& value, Weather& w
       weather.setPressure(static_cast<int16_t>(find->get<float>()));
     find = main.find("humidity");
     if (find != main.end() && find->is_number_unsigned())
-      weather.setHumidity(find->get<int>());
+      weather.setHumidity(static_cast<int8_t>(find->get<int>()));
     foundValidParts = true;
   } // if main object
   find = value.find("wind");
@@ -86,7 +86,7 @@ bool NLohmannJsonOwm::parseSingleWeatherItem(const value_type& value, Weather& w
     const value_type clouds = *find;
     find = clouds.find("all");
     if (find != clouds.end() && find->is_number_integer())
-      weather.setCloudiness(find->get<int>());
+      weather.setCloudiness(static_cast<int8_t>(find->get<int>()));
   } // if clouds object
   find = value.find("rain");
   if (find != value.end() && find->is_object())
